capston-project/requirements.txt
streamlit
python-dotenv
langchain
langchain-groq
groq
supabase
streamlit-extras
